Part - 28 Completed

+++ BloodDonationApp/Controllers/CityTablesController.cs
@@ -17,12 +17,20 @@ namespace BloodDonationApp.Controllers
         // GET: CityTables
         public ActionResult Index()
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             return View(db.CityTables.ToList());
         }
 
         // GET: CityTables/Details/5
         public ActionResult Details(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -38,6 +46,10 @@ namespace BloodDonationApp.Controllers
         // GET: CityTables/Create
         public ActionResult Create()
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var city = new CityTable();
             return View(city);
         }
@@ -49,6 +61,10 @@ namespace BloodDonationApp.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "CityID,City")] CityTable cityTable)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.CityTables.Add(cityTable);
@@ -62,6 +78,10 @@ namespace BloodDonationApp.Controllers
         // GET: CityTables/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -81,6 +101,10 @@ namespace BloodDonationApp.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "CityID,City")] CityTable cityTable)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(cityTable).State = EntityState.Modified;
@@ -93,6 +117,10 @@ namespace BloodDonationApp.Controllers
         // GET: CityTables/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -110,6 +138,10 @@ namespace BloodDonationApp.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             CityTable cityTable = db.CityTables.Find(id);
             db.CityTables.Remove(cityTable);
             db.SaveChanges();
@@ -118,6 +150,7 @@ namespace BloodDonationApp.Controllers
 
         protected override void Dispose(bool disposing)
         {
+            
             if (disposing)
             {
                 db.Dispose();
